+++ Assets/Scripts/CharacterController.cs.meta
@@ -1,2 +1,2 @@
 fileFormatVersion: 2
-guid: 313bb4d86adfdc94d8de1e3b20f83126
+guid: 18a28f18122c15d4187ff11094b90194
