+++ Assets/iniciando-com-unity.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 9f2229ded5e0ef248aa9944158afe12f
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
